fix: map SWI-Prolog `bool` to a 1-byte type to fix x86-64 blob ABI
SWI-Prolog >= 9.3 changed many FLI return types (and some struct fields)
from `int` to C99 `bool` (`_Bool`, 1 byte). The previous workaround
`#define bool int` told bindgen these were 4-byte `int`, which is wrong:
on x86-64 a `_Bool`-returning function only sets AL and leaves the upper
24 bits of EAX undefined, so reading a 4-byte `int` yields garbage in the
high bytes. This made `PL_is_blob`/`PL_get_blob` and other predicates
misfire — e.g. `Term::get_ex::<Blob>()` failed and the type-error term was
left unbound, producing a spurious "tried to raise a var as an exception"
panic. It also mis-sized any struct `bool` field (4 bytes vs 1), corrupting
layouts such as `PL_blob_t`.

AArch64 happens to zero-extend `_Bool` returns, which is why the `int`
mapping worked on Apple Silicon but crashed terminusdb on x86-64 Linux
(store init / `serve --memory`).

Map `bool` to `unsigned char` instead: it matches `_Bool`'s real size and
alignment for both returns and struct fields, while the generated integer
bindings keep the existing `!= 0` / `== 0` comparisons compiling. One
call site (`PL_set_resource_db_mem` result) needed a widening cast.

Verified end-to-end: terminusdb built against this and `serve --memory`
now returns `/api/info` 200 on x86-64 Linux (previously panicked).

+++ swipl/src/init.rs
@@ -96,7 +96,8 @@ pub fn initialize_swipl_with_state(state: &'static [u8]) -> Option<EngineActivat
     // https://www.swi-prolog.org/pldoc/doc_for?object=c(%27PL_set_resource_db_mem%27)
     let result = unsafe { PL_set_resource_db_mem(state.as_ptr(), state.len()) };
 
-    if result != TRUE as i32 {
+    // PL_set_resource_db_mem returns a 1-byte `bool` on SWI-Prolog >= 9.3.
+    if i32::from(result) != TRUE as i32 {
         return None;
     }
